remove usize/isize

+++ container/integer_hash_table.cpp
@@ -32,7 +32,7 @@ template <class K, class V, std::enable_if_t<std::is_integral_v<K>>* = nullptr> 
         static inline constexpr u64 PHI = 11400714819323198485ull;
         static inline const u64 RND = xorshift();
 
-        usize logn, size, full, deleted;
+        int logn, size, full, deleted;
         Manager() : logn(0), size(0), full(0), deleted(0) {}
 
         void fix() {
@@ -42,9 +42,9 @@ template <class K, class V, std::enable_if_t<std::is_integral_v<K>>* = nullptr> 
         }
 
         bool balanced() const { return 2 * (full + deleted) <= size and 8 * full >= size; }
-        usize mask() const { return size - 1; }
+        int mask() const { return size - 1; }
 
-        usize index(const K& k) const {
+        int index(const K& k) const {
             u64 x = static_cast<u64>(k) ^ RND;
             x ^= x >> (64 - logn);
             return (x * PHI) >> (64 - logn);
@@ -54,30 +54,30 @@ template <class K, class V, std::enable_if_t<std::is_integral_v<K>>* = nullptr> 
     Data* data;
     Manager info;
 
-    usize find_key(const K& k, usize i) const {
+    int find_key(const K& k, int i) const {
         while (data[i].ctrl != Ctrl::Empty) {
             if (data[i].ctrl == Ctrl::Full and data[i].slot.pair.first == k) break;
             i = (i + 1) & info.mask();
         }
         return i;
     }
-    usize find_place(usize i) const {
+    int find_place(int i) const {
         while (data[i].ctrl == Ctrl::Full) i = (i + 1) & info.mask();
         return i;
     }
 
-    template <class... Args> void construct(const usize i, Args&&... args) {
+    template <class... Args> void construct(const int i, Args&&... args) {
         new (&data[i].slot.mut_pair) std::pair<K, V>(std::forward<Args>(args)...);
     }
     void resize() {
         Data* old_data = std::exchange(data, nullptr);
-        const usize old_len = info.size;
+        const int old_len = info.size;
         info.fix();
         if (info.size) {
             data = new Data[info.size];
-            for (const usize i : rep(0, old_len)) {
+            for (const int i : rep(0, old_len)) {
                 if (old_data[i].ctrl == Ctrl::Full) {
-                    const usize k = find_place(info.index(old_data[i].slot.pair.first));
+                    const int k = find_place(info.index(old_data[i].slot.pair.first));
                     data[k].ctrl = Ctrl::Full;
                     construct(k, std::move(old_data[i].slot.mut_pair));
                 }
@@ -94,10 +94,10 @@ template <class K, class V, std::enable_if_t<std::is_integral_v<K>>* = nullptr> 
 
     class Iter {
         friend class IntegerHashTable;
-        usize idx;
+        int idx;
         Self* self;
 
-        explicit Iter(const usize i, Self* s) : idx(i - 1), self(s) { next(); }
+        explicit Iter(const int i, Self* s) : idx(i - 1), self(s) { next(); }
 
         void next() {
             while (++idx < self->info.size)
@@ -112,10 +112,10 @@ template <class K, class V, std::enable_if_t<std::is_integral_v<K>>* = nullptr> 
 
     class ConstIter {
         friend class IntegerHashTable;
-        usize idx;
+        int idx;
         const Self* self;
 
-        explicit ConstIter(const usize i, const Self* s) : idx(i - 1), self(s) { next(); }
+        explicit ConstIter(const int i, const Self* s) : idx(i - 1), self(s) { next(); }
 
         void next() {
             while (++idx < self->info.size)
@@ -135,9 +135,9 @@ template <class K, class V, std::enable_if_t<std::is_integral_v<K>>* = nullptr> 
             info.fix();
             if (info.size) {
                 data = new Data[info.size];
-                for (const usize i : rep(0, other.info.size)) {
+                for (const int i : rep(0, other.info.size)) {
                     if (other.data[i].ctrl == Ctrl::Full) {
-                        const usize k = find_place(info.index(other.data[i].slot.pair.first));
+                        const int k = find_place(info.index(other.data[i].slot.pair.first));
                         data[k].ctrl = Ctrl::Full;
                         construct(k, other.data[i].slot.mut_pair);
                     }
@@ -156,14 +156,14 @@ template <class K, class V, std::enable_if_t<std::is_integral_v<K>>* = nullptr> 
     }
 
     template <class... Args> std::pair<V*, bool> insert(const K& key, Args&&... args) {
-        usize idx = -1;
+        int idx = -1;
         if (empty()) {
             info.full += 1;
             resize();
             idx = info.index(key);
         } else {
-            const usize pos = info.index(key);
-            usize i = find_key(key, pos);
+            const int pos = info.index(key);
+            int i = find_key(key, pos);
             if (data[i].ctrl == Ctrl::Full) return std::make_pair(&data[i].slot.pair.second, false);
             i = find_place(pos);
             info.full += 1;
@@ -182,7 +182,7 @@ template <class K, class V, std::enable_if_t<std::is_integral_v<K>>* = nullptr> 
 
     bool erase(const K& key) {
         if (empty()) return false;
-        const usize i = find_key(key, info.index(key));
+        const int i = find_key(key, info.index(key));
         if (data[i].ctrl == Ctrl::Full) {
             info.full -= 1;
             info.deleted += 1;
@@ -196,7 +196,7 @@ template <class K, class V, std::enable_if_t<std::is_integral_v<K>>* = nullptr> 
 
     V* find(const K& key) const {
         if (empty()) return nullptr;
-        const usize i = find_key(key, info.index(key));
+        const int i = find_key(key, info.index(key));
         return data[i].ctrl == Ctrl::Full ? &data[i].slot.pair.second : nullptr;
     }
 
@@ -208,7 +208,7 @@ template <class K, class V, std::enable_if_t<std::is_integral_v<K>>* = nullptr> 
         info = Manager();
     }
 
-    usize size() const { return info.full; }
+    int size() const { return info.full; }
     bool empty() const { return size() == 0; }
     V& operator[](const K& key) { return *insert(key).first; }
